Rename to EMA for consitency with other implementations
